requirements.txt
asgiref==3.8.1
awsebcli==3.23.3
blessed==1.21.0
botocore==1.38.10
cement==2.10.14
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==4.2#3.2.5
django-environ==0.12.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pathspec==0.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.13.2
urllib3==1.26.20
wcwidth==0.2.13
psycopg2-binary==2.9.10
python-dotenv==1.1.0
django-adminplus==0.6
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-multiselectfield==0.1.13
#Django==3.2.5
#psycopg2-binary==2.9.10
#python-dotenv==1.1.0
#django-adminplus==0.6
#djangorestframework==3.16.0
# django-cors-headers==4.7.0
#django-multiselectfield==0.1.13